backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
python-multipart==0.0.6

# LangChain
langchain==0.1.6
langchain-openai==0.0.5
langchain-community==0.0.19

# Vector Store
chromadb==0.4.22

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
unstructured==0.12.0

# Embeddings
sentence-transformers==2.3.1
openai==1.12.0

# Utils
pydantic==2.5.3
pydantic-settings==2.1.0
